SmrPlanetType: fix PHPStan issue
PHPStan was reporting the following error:

> Access to undefined constant static(SmrPlanetType)::STRUCTURES.

This is a deficiency in PHPStan, because the code functioned properly
thanks to "late static binding", where using `static::` refers to the
class that was instantiated (a child) rather than the class in which
it was called (the parent) -- in contrast to `self::`, which always
references the current class.

Since SmrPlanetType is defined as abstract, it cannot be instantiated
directly, which means that the absence of a `STRUCTURES` constant is
not problematic, so long as all of its children define it. However,
this is a subtlety PHPStan does not yet understand, as evidenced by
issues like https://github.com/phpstan/phpstan/issues/4061.

That said, it is not manifestly obvious that child classes must define
a `STRUCTURES` constant. Ideally, we would declare it as `abstract` in
the parent, but there is no such thing as abstract constants in PHP.

We also don't want to add `const STRUCTURES = [];` in the parent class
(just to satisfy PHPStan), because then omitting it from a child class
would not cause an error, and it would result in using the wrong data.

Therefore, we add a wrapper function `getStructureData`, which is
abstract in the parent class, and simply returns `self::STRUCTURES` in
the child classes. It's technically unnecessary abstraction, but it
solves the PHPStan issue while at the same time making the interface
more robust.