docker-compose.yml: add smr-builder YAML anchor
Many of our services were building their own image, but were able to
avoid doing any extra work thanks to the Docker layer cache. Basically,
this was the same image but with many different names.

In some sense this was nice because it decoupled the services. Building
for one service would not modify the container for other services.
This was a double-edged sword, however, because often we wanted to keep
the images in sync, but it was easy to forget the build step for some
of the services.

By specifying an `image` field in addition to `build`, we are telling
Docker to build the image with the name indicated by `image` (rather
than giving it an automatic name based on the directory/service name).
Now all the services that specify `image` will use the updated image,
regardless of which service it was originally built for.

The common `build` and `image` fields are refactored into a YAML anchor
called `smr-builder` for easy reuse in the services that need it.

The choice of docker-compose to overload `image` for local and remote
images is a poor design, because now `docker-compose pull` will warn
about all the services that use a local image. For this usage, `build`
should have had its own `image` sub-field.