shared.css: improve Local Map port good spacing
When we removed whitespace stripping from Template in 83655c40a2,
the spaces between port good images on the Local Map caused them to
overflow onto a second line. We set the font size in this div to 0
to recover the original compact spacing (and now in _most_ cases the
goods should not spill onto a second line).