Remove htmliseMessage function
This was only used in one place (box_view.php), and can easily be
replaced with standard PHP functions.

We don't really need to bbify when displaying admin messages. The
raw BBCode should give us all the info we need, but we could mess
with the `$overrideGameID` global if we really wanted to display
nicely bbified text.