Add .scrutinizer.yml
This configuration used to be stored online at Scrutinizer-CI, but it
was difficult to track when something needed to change (such as PHP
version or directory structure). The configuration included here is
merged with the online configuration.

See https://scrutinizer-ci.com/docs/configuration/cascade.

The default image is Ubuntu 14.04(!), so we explicitly specify 18.04
(their only other choice presently). This is necessary because OpenSSL
in 14.04 is limited to 1.0.1f, but PHP 8.1+ requires OpenSSL 1.0.2+.

See https://prinsfrank.nl/2022/02/04/Configuring-PHP8.1-on-scrutinizer.