Add Database::insert method
Manual construction of SQL statements is fragile and dangerous, since
there is no way to validate the resulting string apart from executing
the query.

We replace all basic INSERT statements with calls to Database::insert,
with structured data passed as an argument. (In the future, we should
apply this strategy to other statement types, but UPDATE statements,
as an example, are often more complex.)

This design has some very important benefits:

1. The key-value pairs are directly (and visually) associated, so it
   is much harder to pass the wrong value to a column.

2. It is impossible to create a string with a SQL syntax error, since
   all of the formatting (commas, parentheses, etc.) is automatic.

3. If we can unify the way that we escape content in the future, then
   we can construct the `$fields` array with raw values, and perform
   the escaping irrespective of the value types. (Though this could
   possibly degrade specificity, if we're not validating the types
   against a schema for the given table, for example.)

4. Just generally it is _much_ easier to read.