RouteGenerator: fix corner case with empty routes
Due to the way that `$lastGoodIsNothing` was being updated, it could
cause us to incorrectly allow "two nothings in a row" for successive
elements of `$forwardRoutes`.

This is likely of little consequence, since they would just be extra
terrible routes in the list, but it's good to rectify the logic.