Refactor SmrAccount constructors
If an account cannot be found, it now throws an exception instead of
returning null. This is important for ensuring that the `getAccount*`
functions all return a single type. It also makes it consistent with
the other fundamental object classes (SmrPlayer, SmrAlliance, etc.).

As a result of this change, all places that check for null returns
needed to be modified to do a try-catch instead.