Session: simplify session ID creation (#1267)
We do not need to use a cryptographically strong ID for the session,
but it doesn't hurt. Our `random_string` function is perfect for this
job, and even allows us to explicitly specify the length.

This change is motivated by the poor properties of `uniqid` (it is
based on the system time, which is why we were generating a prefix
string from a random int), which were further degraded by taking the
md5 hash of the result.

In the tests that compute an SN, we see a different random string
because we're no longer calling `rand()` inside the Session ctor.