Fix code smell UnneededElse
PSR2R.ControlStructures.UnneededElse

In general, if/else clauses that create exit points (continue, break,
return, etc.) are unneeded because the code that follows will not be
reachable. The extra indentation caused by this code smell can also
be harmful to code readability (since deep nesting is hard to read).

When a condition clearly indicates a departure from the "normal"
execution path, then:

  if (condition) {
    return foo
  }
  return bar

matches the asymmetry of the logic better than:

  if (condition) {
    return foo
  } else {
    return bar
  }

The former also has the benefit of reducing indentation levels, and
thus (usually) making the code easier to read.

At first, I was skeptical that this was a good code sniff, but after
reassessing each case carefully, I realized that it was more often the
case that fixing this code smell improved its readability. Of course,
there will be exceptions, especially when there is more symmetry in
the various exit points.

In some cases, we replace the if/else clauses with a match expression
or a ternary operator, both of which have the added benefit of being
a single expression (and thus a single exit point).