Avoid incomplete parameter coverage in docstring
It seems that phpcs can get confused if we only partially document a
function interface. Doing so may cause it to try to match an argument
that is undocumented with one of the docstring arguments, and thus
potentially emit a type mismatch error.

Therefore, if we document _any_ function arguments, we should do it
for all of the arguments.