Refactor pluralise function
Include the amount in the returned value to simplify the way it is
used in a majority of cases.

The only times when we don't want to include the value, it is because
we are using some special HTML formatting of the amount. This is why
the function now has a `$includeAmount` boolean argument.

We remove the "is -> are" functionality, since places where this was
used could be reworded in a sensible way.