Globally trim strings input by the user
In many cases where we use `Request::get`, we trim the returned value
manually. This is because leading and trailing whitespace from a form
input is almost always accidental and/or meaningless.

For example, I was copy-pasting the alliance name when creating NPCs,
and because we were _not_ trimming that form input, I ended up creating
a second alliance due to a trailing whitespace artifact from the copy.

Now we will _always_ trim form inputs in `Request::get`, and so we no
longer need to call `trim` explicitly on its returned value.