Unigen: fix 1-sector width/height galaxies
While topologically it makes sense for a sector to connect to itself
in a galaxy with a width or height of 1 sector, it is probably going
to be confusing to the player, and presently causes errors in the
universe generator.

SmrSector::enableLink will now silently do nothing if the sector would
link to itself.