docker-compose.yml: use entrypoint for test services
The `entrypoint` field defines the base command that should be run,
even if commandline arguments are passed. This is very useful if we
want to run the test services with a more specific option.

For example, to run phpstan on an individual file, we previously would
need to run:

  dc run --rm phpstan vendor/bin/phpstan --memory-limit=4G --ansi analyse path/to/file.php

With `entrypoint` defined, we can now run:

  dc run --rm phpstan path/to/file.php

And similarly, with composer scripts (due to the trailing `--`):

  composer phpstan path/to/file.php