Avoid throwing in autoload function
Our custom autolader `get_class_loc` would _always_ try to load the
fallback file, even if it didn't exist. This interacted poorly with
the `class_exists` function (with autoload=true) in cases where we
were not expecting the class to exist.

In particular, this broke PHPUnit testsuite "name" attribute (in the
phpunit.xml configuration), because it's an arbitrary name, and yet
PHPUnit still checks to see if it is an autoloadable class for some
reason. With this fixed, we no longer need to explicitly pass the
`test` argument to the PHPUnit invocation, which was overriding the
testsuite config.

When `get_class_loc` doesn't find a file associated with the class,
we will now get a "class not found" error message instead of a file
loading failure (assuming it was not loaded by the PSR-4 autoloaders).

Note that `get_class_loc` now returns a bool. This isn't strictly
necessary for an autoload function, but it may be useful if we ever
find the need to test whether it actually loaded a file.