phpunit: move colors option to config file
Set the use of color output in the config file instead of on the
commandline. While `colors="true"` is equivalent to `--colors=auto`,
it still seems to output colors as expected.

This may also fix the issue with the GitHub Actions output being
cluttered with escape characters (see b8591d62d3).