Change NPC_SCRIPT from a const to a DI variable
Setting NPC_SCRIPT as a PHP constant made it difficult to test classes
that depended on it. This is because PHPUnit has no mechanism to change
the value of a PHP constant, apart from `@runInSeparateProcess`, and
then setting the constant to a different value in that process.

Unfortunately, `@runInSeparateProcess` interacted poorly with PHP-DI
in a way that I don't completely understand. When running tests using
this decorator, the following error would be triggered:

  PHPUnit\Framework\Exception: PHP Fatal error:  Uncaught Error: Using $this when not in object context in /smr/vendor/php-di/php-di/src/Compiler/Template.php:4
  Stack trace:
  #0 Standard input code(1154): require_once()
  #1 {main}
    thrown in /smr/vendor/php-di/php-di/src/Compiler/Template.php on line 4
  Fatal error: Uncaught Error: Using $this when not in object context in /smr/vendor/php-di/php-di/src/Compiler/Template.php:4
  Stack trace:
  #0 Standard input code(1154): require_once()
  #1 {main}
    thrown in /smr/vendor/php-di/php-di/src/Compiler/Template.php on line 4

This error always occurred when running the associated test files
directly, but in an upcoming class `SectorLock`, it also occurred when
running the entire test suite, which is obviously a blocking issue.

To fix this, we convert the PHP constant `NPC_SCRIPT` into a PHP-DI
variable with the same name. This allows us to change its value more
easily, and particularly at runtime instead of compile time. In some
sense this is an abuse of the DI container, since we're not actually
using it for DI -- rather just for global variable storage. Oh well.