SmrShipType.php: rewrite power calculations
No functionality changes.

This is a more declarative way to map hardpoints to power allowance.
Previously, we were using a somewhat esoteric aspect of the switch
statement, which is that without an explicit `break` statement, all
subsequent branches are triggered. While this is perhaps more elegant
from a numerical perspective, the code is too opaque.

Fixes code smell:
PSR2.ControlStructures.SwitchDeclaration.TerminatingComment