Rename SmrPlanet::accuracy to getTurretAccuracy
Clarify that this method is only for informational display purposes,
and is not used directly in any combat calculations.

This also fixes a potential error for any planet type that has both
mounted weapons and turrets (though no such planet type exists yet).
It would have taken the accuracy of the mounted weapon instead of
the turret, since the turrets are appended to the weapon array.
Now, we just query the turret accuracy explicitly.