Handle redirect hooks consistently
There are a few special hooks that force the player to view a specific
page that they did not trigger: game end, death, newbie turn warning.

To avoid infinite loops, we need a mechanism to skip these hooks when
we're trying to process these pages. We do this with a `skipRedirect`
argument to `Page::create`, which adds a special var `_skip_redirect`
into the container that can be queried with the `skipRedirect` method.

This takes the place of checking if the page to process is the same as
the redirect hook page (which was notably missing for the "game end"
hook, causing an infinite loop in this scenario), as well as the more
ad-hoc `override_death` var.

Note: The `skipRedirect` mechanism skips _all_ redirect hooks, whereas
we were previously targeting a specific hook to skip. This seems fine
with our current hooks, but it's not clear if we will need a more fine
tuned skipping mechanism if we add different hooks in the future.

Note: The `Globals::getPodScreenHREF` method has been removed, and we
replace it with `Globals::getCurrentSectorHREF` in the attack display
pages where the player has died. It seems better to just let the
redirect hook catch this case, instead of explicitly calling the
death processing page here (and we would have needed to add the
`skipRedirect` to `Globals::getPodScreenHREF`).