Fix multiple lock error after death
With the more strict sector locking checks added with the `SectorLock`
class in 54e83b9d00, we were getting "multiple sector lock" errors on
player death. This was because they were in one sector during the
attack processing (where they died), and then in another sector when
forwarded to the display page (their HQ).

The fix is fairly straightforward: once processing of the attack has
finished, if the player is dead, flush all the caches and release the
lock (and then re-acquire the lock in the new sector to avoid reloading
the session var, as usual). Note that we could update the session here
as well, but since we are skipping the session var reload, this would
just be unnecessary database work.

We refactor the cache flushing into a function `saveAllAndReleaseLock`
(which might be better as a SectorLock method?). Some `class_exists`
checks have been added as a minor optimization in case we haven't yet
loaded any of the Smr{Force,Port,Planet} classes (and thus their caches
would of course be empty anyway).