Display all eligible fighting ships on Examine
This will help players make a more informed decision about whether or
not to engage if the number of players exceeds the fleet combat size,
especially in conjunction with the new ship totals.

Limiting the ships in the Examine page doesn't even make sense since
the subset of ships chosen by `limitFightingTraders` is randomized.
Therefore, the Examine list would be different than the players who
are actually selected for that round of combat.

We also make some nice simplifications to `getFightingTraders`, with
no functionality changes (apart from the added `allEligible` option).