Speed up tests that modify the database
The goal is to have the database return to its initial state after
every test, so that tests are independent of each other.

The way we were previously achieving this was by storing a list of
tables that start empty and truncating all those tables after every
test, since we (happen to) only modify tables that start empty.
This is the slowest method (apart from re-initializing the database
after every test) because we currently have ~120 empty tables, but
most test only modify a handful of them. This incurs a lot of overhead
cost, even with empty tables. With this method, the tests take ~23
seconds.

A faster way to do this is to count the number of rows of each table,
and then only truncate tables that change from zero to non-zero rows.
I suspect this is faster because counting rows is done in a single
query, whereas each `TRUNCATE` must be done in a separate query (for
whatever reason). This still involves a fairly expensive query after
each test, but brings the total time to ~17 seconds.

The fastest method is to explicitly specify the tables that need to be
reset after every test for each class, and only truncate those tables.
This is the method that we now use, and takes ~12 seconds. We add a
sanity check after each test class to make sure that all tables are
properly truncated, and we fail if any have changed size. (This helps
us identify which, if any, tables need to be added to the list of
tables to truncate for any given class.) This still requires us to
initialize a list of table sizes once for the entire test suite.

Counting table rows is actually pretty complicated, since the values
that are queried from the `information_schema` are not accurate until
we run an `ANALYZE TABLE` query on every table. This turns out to be
a major expense (though less than blindly truncating every table), so
as an alternative to counting table rows, we can use `CHECKSUM TABLE`
queries, which is always accurate (and can even be optimized for our
MyISAM tables, which can be configured to properly cache their
checksums; though this is a complicated optimization that we probably
do not need, and we may want to be using InnoDB ultimately anyway).
Since we only care about empty vs. non-empty, we just look for tables
whose checksum is not zero. This significantly reduces test overhead,
and the whole test suite only takes ~8 seconds.

A few special notes:

* Initial table names are not stored for faster `CHECKSUM` queries
  since we want to be aware if any tables are created or dropped.

* We use the special `@before(Class)` and `@after(Class)` annotations
  on the `BaseIntegrationSpec` methods so that we don't need to worry
  about accidentally overriding the setup/teardown methods in child
  classes (without calling `parent::*`, which is no longer needed).

* The `smr_live_test` database name is used explicitly in the base
  class to avoid accidentally using the live database (e.g. if we
  forget to pass the `test/env` file to docker-compose.

* The `SectorLockTest` can now inherit from `BaseIntegrationSpec` and
  its special table truncating logic can be removed.