Enforce uniqueness of account/player in database
The following properties now have a UNIQUE constraint on them in the
database:

* Account email
* Hall of Fame name
* Player ID (per game)
* Player name (per game)

These need to be unique for the code to work properly, and it was
previously only enforced (inconsistently) at the PHP level.

It should also result in slightly faster database lookup (at the cost
of increased database size) for the `get{Account,Player}By*` methods,
since UNIQUE adds an index on these columns.

More consistent and user-friendly handling of this uniqueness within
the PHP code will be forthcoming.