SmrPort: add/remove goods on update
Add a new attribute `goodExistenceChanged` that tracks which goods, if
any, were added or removed from the port since the last query/update.
This allows us to update the database only in the `update` method, and
avoid making database changes in `{add,remove}PortGood`.

This is operating under the principle that most methods should not
touch the database, if possible.

Also fix an issue where the `goodAmountsChanged` attribute was not
properly being reset after updating good amounts in the database. The
result of such a bug is that the UPDATE queries would continue to be
run each time `update` was called.