phpcs.xml: disable UnaryOperatorSpacing.WrongSpace
In PHP, both `$a =& $b` and `$a = &$b` do the same thing, but only the
former clearly indicates the unexpected way that reference assignment
actually behaves.

In many other languages, assigning a reference to a variable `$a = &$b`
will make the type of `$a` a reference, and assigning something else to
`$a` will overwrite this reference. However, in PHP, assigning to `$a`
_changes_ `$b`, because reference assignment creates a _link_ between
the two variables. In this way, `=&` is not equivalent to the separate
operations of assignment `=` and referencing `&`, which makes the use
of `$a = &$b` misleading.

For this reason, I prefer to use `=&` because it more clearly indicates
how PHP reference assignment actually behaves. Unfortunately, this is a
syntactical outlier, and as such is generally warned about in style
checks. While we can (and should) limit how much we use assignment by
reference, we will likely never get rid of it completely, and therefore
need to suppress the PSR2R.WhiteSpace.UnaryOperatorSpacing.WrongSpace
code sniff.

For further discussion, see:
https://stackoverflow.com/questions/1768343/reference-assignment-operator-in-php