Add unit test for MultiplePortRoute
This includes a 3-way port test, since it is the most complicated
usage of the class.

When type hints were added in c50845acc, I didn't appreciate that the
MultiplePortRoute class used recursion to extend the forward/return
route design to more than 2 ports.

The "forward" route is the one that gets nested, and the "return"
route is always a OneWayRoute. At the end of the chain, the "forward"
route will also be a OneWayRoute.

The typehints are now fixed. Closes #1268.