RouteGenerator: fix remnants of buy/sell ambiguity
There was always some confusion about what "buy" and "sell" meant in
the code (is it from the perspective of the port or the player?).

When it was decided to try to do everything from the perspective of
the player, Page's route generator code was the most impacted because
because it was written from the other perspective (as I learned from
Page much later).

This was sorted enough for the RouteGenerator to output correct
routes and multipliers (verified independently), but it always had a
weird quirk where the order of the routes seemed to be backwards.
This is finally fixed in a comprehensive way that should be fully
consistent with our player-centric wording.

Old:

13 (Human) buy Nothing for 0x to sell at 1 (Alskant) for 0x (Distance: 4)
7 (Neutral) buy Food for 2x to sell at 13 (Human) for 2x (Distance: 2)
1 (Alskant) buy Precious Metals for 2x to sell at 7 (Neutral) for 2x (Distance: 2)

New:

1 (Alskant) buy Precious Metals for 2x to sell at 7 (Neutral) for 2x (Distance: 2)
7 (Neutral) buy Food for 2x to sell at 13 (Human) for 2x (Distance: 2)
13 (Human) buy Nothing for 0x to sell at 1 (Alskant) for 0x (Distance: 4)