RouteGenerator: remove generateOneWayRoutes
With a small tweak, `generateMultiPortRoutes` can be made to support
one-way routes. Just like in `generateOneWayRoutes`, the return route
is an empty route (i.e. GOODS_NOTHING), so GOODS_NOTHING has to be
turned on for one-way routes to generate (otherwise no routes will
be returned).

For the sake of completeness, we turn on GOODS_NOTHING in the UniGen
map checker, since it's possible (though not likely) for a one-way
route to have the best profit and/or exp.