Route: add method getPortSectorIDs
This is a complementary function to `containsPort`, but instead returns
a list of sector IDs for all the ports that constitute the route.

We need this functionality to properly handle listing the port sectors
in the NPC script, since `get{Buy,Sell}SectorId` do not exist on
MultiplePortRoute objects. Technically it is only needed if we use
a 3+ port route, but the issue is raised by static analysis regardless.