Fix NPC script to properly handle 3+ port routes
Rewrite the entire trade route logic in the NPC script to properly
support 3+ port routes. Specifically, we were calling methods that
only exist for OneWayRoutes on objects that can be MultiplePortRoutes
(but only when there are 3+ ports).

We add a method Route::getOneWayRoutes, which returns an ordered list
of all the OneWayRoutes that constitute the MultiplePortRoutes (which
requires recursing through the Route tree to extract the leaves). We
then can loop over these OneWayRoutes in the NPC script to find the
specific route that we need to interact with.

An attempt has been made to make the logic in the NPC script more
intelligible, but it's still a bit too complicated and subtle.