Plotter: filter port to port distances by race
If the RouteGenerator will only be using ports from a subset of races,
then it is wasteful to calculate distances for ports of other races,
since they will go unused.

This is a minor speed optimization, most notable for the NPC script,
which dramatically limits the races that it trades with.