Allow NPCs to trade in any Racial/Neutral galaxy
Because the Plotter and RouteGenerator require a contiguous sector
range, we can only select the galaxies until the first Planet galaxy
is encountered. If no galaxies are selected with this method (i.e. if
the first galaxy is a Planet galaxy), then we will just use the NPC's
current galaxy only.

If this causes memory exhaustion, then we may need to rethink this
feature (or optimize the route generator).