SmrEnhancedWeaponEvent: handle unlikely duplicates
If we happen to select the same configuration (sector, loc, weapon)
twice in a row, just replace the existing event. This is an extremely
unlikely scenario, except perhaps in a game that has a limited number
of weapon shops for some reason. This prevents a database error.

We aren't losing much by replacing the existing event, since at most
it only misses out on the grace period where 2 events can exist at
the same time (and only the newer is advertised by the bartender
anyway).