Fix account edit if account already has exception
The `exception_add` field is not present if the account already has
an exception, which was causing any other editing of such accounts to
fail. We give it a default value of the empty string, which will skip
adding an exception.

Also make the "Add An Exception" text a placeholder instead of a value
so that we don't have to make a special case for it.