SmrPlayer: add changePlayerName method
This method refactors the logic for changing a player name that was
being done redundantly in both the "Edit Account" admin tool and in
"Preferences".

Also, "Edit Account" was missing the ability to change the case of the
player name, which was added to "Preferences" in 177401a8e0f, as well
as the basic sanity checks provided by the DisplayNameValidator.

We add this new method instead of putting the logic into the method
SmrPlayer::setPlayerName so as to retain the lower-level API pattern
of setters/getters.