Consolidate config.specific.php files
Having all customizable configuration options in a single file
simplifies configuration management. It does mean that a handful of
constants are defined for services that don't need them, but I think
this is a trivial price to pay to ensure that bootstrapping is
consistent across all services. Regardless of where we use any of
these constants, we can now be sure it is defined.

This also helps simplify static analysis (esp. PHPStan), because it
does not run the code (except for bootstrap.php), and having different
bootstrapping requirements for different files is not maintainable.