bootstrap.php: avoid initializing resources during logging
In `logException`, we don't want to initialize resources if they have
not already been initialized (i.e. Session and SectorLock instances).
We were using `class_exists` to do this, but it turns out that the DI
container loads these classes when the container is initialized to do
autowiring.

With the introspection provided by the new ResettableContainer, we can
finally skip unnecessary resource creation. This is especially useful
in the following scenarios:

* When we don't have a session (i.e. Discord/IRC services).

* When the error occurs during startup before the DI container is
  initialized.

* When the error occurs while initializing these resources (to avoid
  throwing while logging).