docker-compose.yml: allow ctrl+c discord service
This allows us to run `docker-compose run --rm discord` and kill the
interactive process with ctrl+c.

I'm not entirely sure if this cleans up the connection to Discord
properly, but it makes development a lot easier. (In production, we
daemonize this process anyway.)

See 4a0006cd727 for more details.