Fix NPC dumping goods when routes are drained
The fragile logic in npc.php for trading caused NPCs to dump cargo if
they arrived to find a drained route, but then they would just continue
with their route and buy goods instead of picking a new route. They
would progressively lose cash this way, and would eventually fail by
running out of credits!

This is fixed by being much more prescriptive about what an NPC should
be doing when on a trade route. We use the new class `RouteIterator` to
define exactly which action should be taken and when. The resulting
logic in npc.php is more robust and _much_ simpler.

Tangential changes:

* Changed where we increment the action counter so that it does not
  depend on a particular Exception being thrown.

* Changed how we pick a random route from the available routes, using
  `shuffle` and `array_pop` to simplify the logic.