Refactor HOF code
Treat HOF types as they exist in the database as first-class entities.
These are simple strings, with semicolon-delimited elements:

  'a:b:c'

This replaces the multi-dimensional array storage used previously:

  ['a']['b']['c']

The latter was simply too complicated, and involved lots of recursive,
opaque code to map from the database HOF types.

The beneficiaries of this change are:

* SmrPlayer::getHOFData
* SmrPlayer::getHOFVis
* SmrPlayer::getHOF
* SmrPlayer::setHOF
* SmrAccount::getHOFData
* SmrAccount::getHOF

One consequence of this change is that it is now natural to support
HOF types that are substrings of other types, e.g. `a:b` and `a:b:c`.
Special care has to be taken when display such types (though none
presently exist), since there is an ambiguity between displaying the
ranks for the type `a:b` and displaying the subcategory list `a:b:*`.

The simplifications made to the HallOfFame class and the engine files
`hall_of_fame_*.php` should make this ambiguity manageable, though it
does result in a slightly inconsistent interface (i.e. you'll go to
the rankings page for `a:b` if you click on the category link, even if
there are subcategories `a:b:*`, whereas normally it will take you to
another category page).

Additional changes:

* SmrPlayer::getHOFVis now returns its cached data, so we can use this
  instead of manually (and repeatedly) querying the `hof_visibility`
  table.

* The special case HOF types for donations and user score are added
  directly to the HOFVis cached data so that we don't need to special
  case them when using the HOFVis array.

* When displaying HOF categories, HOF ranks are now only computed for
  complete HOF types. Previously, partial types were being attempted,
  resulting in many unnecessary queries.