DatabaseRecord: more strict validation for int/float
If we are expecting a certain type back from the database, it better
well be that type, otherwise we should (and now do) throw an error.

The explicit casting was dangerous, since it could hide discrepancies
between the SQL and the PHP expectations. In particular, I found an
example where it was converting `null` to 0, which may not always be
the appropriate conversion.

Note that we could have done this for bool as well, but the
`FILTER_VALIDATE_BOOL` filter accepts too wide a range of inputs
(e.g. the empty string maps to false) and is more suitable for HTTP
request data.