Fix SQL errors due to null -> int casting
If a table is empty and we attempt to use an aggregating SQL function
on it (e.g. MIN, MAX, SUM), then the result is NULL. In most cases, we
were expecting an integer result without checking for a null value.
Previously, this only worked due to a bug in `DatabaseRecord::getInt`
(now fixed) that would implicitly convert null values to 0.

We use an IFNULL function to define the default value if there are no
matching records. This allows us to ensure that we have an int value
in the resulting field.

This type of usage often checked if there were any records, but there
is _always_ a record returned when aggregating functions are used, so
the logic in these places has been modified accordingly.

Some columns have been adjusted to return 0 as the "next ID" when no
records are found. There is no reason _not_ to have a 0-based ID,
except in special cases where the ID is a public attribute (such as a
player ID or alliance ID). The following columns will now start with
0 instead of 1:

* album_has_comments.comment_id
* alliance_thread.thread_id
* alliance_thread.reply_id
* bar_tender.message_id
* changelog.changelog_id
* galactic_post_article.article_id
* message_notify.notify_id

These would probably all benefit from simply being an AUTO_INCREMENT
column instead of manually calculating IDs (which is also vulnerable
to database collisions if we're not locking carefully).

Note that we can also use `GROUP BY` on a column that is restricted to
a single value by `WHERE` to change the behavior of aggregating
functions, and return no records instead of null. This is useful in a
few specific places, but should not be overused.

Unrelated changes:

We also remove the `thread_replies` limit on the query where we select
all the replies to an alliance topic, since we never actually impose a
limit that is less than the number of replies.