bar_gambling_processing.php: fix phpstan warning
> Variable $bet might not be defined.

We can always use `Request::getVarInt` to get the current bet, because
it will always be defined in `$var` (except the first new game, when
we want to get it as a request variable).

Semi-relatedly, rename the value of `$do` "nothing" to "new game",
since the latter is what is actually represents.