Remove useless LIMIT 1 in SQL queries
We were including `LIMIT 1` in many SQL queries that can only return
(and are _designed_ to only return) one record. In my opinion, using
`LIMIT 1` in this scenario is an anti-pattern, because it applies an
artificial constraint on the query.

There are some arguments for including it: specious ones regarding
performance (it is not faster), and ones that recommend it in the
name of safety (though I think this is approaching safety from the
wrong direction). Instead, I think `LIMIT 1` _at best_ will hide the
fact that you have an improper query if it is supposed to return one
record but would have returned more (had it not been for `LIMIT 1`).
The one query you do get will be silently erroneous or misleading.