chess.php: fix phpstan warning
> Variable $move might not be defined.

The new version is a bit clunkier, and I'm not sure phpstan is correct
that `$move` is not guaranteed to be defined, but the original code
really lacked transparency, so I don't mind changing it.

Also replace `stripos` with `str_starts_with`. The latter is not case-
insensitive, but I'm pretty confident that "bestmove" is expected to
be all lowercase, though I can't find any documentation about "Seg".