npc.php: add NpcActor class
This refactors all of the logic from the main NPC script driver into
a separate class to help manage all the variables that apply only to a
single NPC. This allows us to simply replace the instance of NpcActor
to reset all of these variables when we change NPCs.

The exception handling has also been refactored a bit so that the main
action function of NpcActor (`getNextAction`) does not concern itself
at all with the handling. That is now the role of the `npcDriver`
function.