npc.php: do not modify $_REQUEST superglobal
Some processing pages that are called by the NPC script do require data
from requests, but we can put that data inside the container instead,
and use `Request::getVar` instead of just `Request::get` in the page
processing.

This allows us to simplify some of the bookkeeping of the NPC script.